submissions/guilherme-pardin/solution/requirements.txt
streamlit
pandas
anthropic
google-genai
openai
plotly
